% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_log_ratio.R
\name{compute_log_ratio}
\alias{compute_log_ratio}
\title{Function to compute log ratio between the normalising constant and the estimates}
\usage{
compute_log_ratio(Z, fkf.obj)
}
\arguments{
\item{Z}{Estimate results given by the algorithm}

\item{fkf.obj}{Normalising constant of the Gaussian distributions computed by FKF}
}
\value{
Log ratio between real values and estimates
}
\description{
Function to compute log ratio between the normalising constant and the estimates
}
