% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_quasi_online_pf.R
\name{run_quasi_online_pf}
\alias{run_quasi_online_pf}
\title{Function to run the whole iAPF processes and generate final results}
\usage{
run_quasi_online_pf(model, data, lag, Napf, N)
}
\arguments{
\item{model}{Model information}

\item{data}{Observations}

\item{lag}{Length of time intervals}

\item{Napf}{Number of particles used in the smoothing process}

\item{N}{Number of particles used in the filtering process}
}
\value{
A list contains
X is the particles
w is the weights of particles
logZ is the estimated normalising constant
Xs is the smoothing particles
log_ratio is the log_ratio between estimates and real normalising constant
dist is the distance between real estimates and smoothing distribution
}
\description{
Function to run the whole iAPF processes and generate final results
}
